requirements.txt
flask==3.0.0
flask-cors==4.0.0
flasgger==0.9.7.1
mysql-connector-python==8.2.0